Exported out sidebar comp
