+++ client/src/components/dashboard/RealmsSideBar.tsx
@@ -1,0 +1,61 @@
+import { FaArrowRightFromBracket, FaComments, FaPlus, FaUser } from "react-icons/fa6"
+import Modal from "../ui/Modal"
+import CreateRealmForm from "./CreateRealmForm"
+import { useUser } from "@auth0/nextjs-auth0/client";
+import { FormEvent, useRef } from "react";
+
+const RealmsSideBar = () => {
+	const { user } = useUser();
+	const username = user?.name;
+
+	const dialog = useRef<HTMLDialogElement>(null);
+
+	const openModal = (e: FormEvent) => {
+		e.stopPropagation();
+		if (dialog.current && !dialog.current.open) {
+			dialog.current.showModal();
+		}
+	}
+	const closeModal = (e: FormEvent) => {
+		e.stopPropagation();
+		if (dialog.current) {
+			dialog.current.close();
+		}
+	}
+
+	return (
+		<div className="h-[100%] bg-slate-900 w-[100px] pt-3">
+			{/* Profile */}
+			<div className="flex flex-col w-[100%]">
+				<FaUser size="3em" className="text-slate-500 self-center" />
+				<p
+					className="self-center text-sm max-w-[85%] overflow-hidden text-ellipsis whitespace-nowrap no-underline"
+					title={username ?? ""}>
+					{username ?? ""}
+				</p>
+			</div>
+			<div className="flex flex-col bg-slate-800 py-3">
+				<FaComments size="3em" className="self-center text-slate-500" />
+			</div>
+			<div className="flex flex-col bg-slate-900 py-3">
+				<a
+					className="hover:cursor-pointer self-center"
+					href="/api/auth/logout"
+					title="Log out"
+				><FaArrowRightFromBracket size="3em"
+					className="bg-transparent text-slate-500" />
+				</a>
+			</div>
+			<div className="flex flex-col bg-slate-900 py-3 hover:cursor-pointer"
+				onClick={openModal} >
+				<FaPlus size="3em" className="self-center text-slate-500" />
+			</div>
+			<Modal ref={dialog}>
+				<CreateRealmForm dialog={dialog} />
+				<button type="button" onClick={closeModal}>Close</button>
+			</Modal>
+		</div>
+	)
+}
+
+export default RealmsSideBar;
